+++ changelog.md
@@ -5,3 +5,4 @@
 1. Setup base application using web
 2. Added Amazon dependency
 3. Added S3 bucket health check
+4. File upload and download